(#196)feat: RepositoryUtil 클래스 생성했습니다.
