+++ src/main/java/com/coverflow/global/util/RepositoryUtil.java
@@ -1,0 +1,34 @@
+package com.coverflow.global.util;
+
+import com.querydsl.core.types.Order;
+import com.querydsl.core.types.OrderSpecifier;
+import com.querydsl.core.types.dsl.EntityPathBase;
+import com.querydsl.core.types.dsl.PathBuilder;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+
+public class RepositoryUtil {
+
+    private RepositoryUtil() {
+    }
+
+    public static <T> OrderSpecifier[] makeOrderSpecifiers(final EntityPathBase<T> qClass, final Pageable pageable) {
+        return pageable.getSort()
+                .stream()
+                .map(sort -> toOrderSpecifier(qClass, sort))
+                .toList().toArray(OrderSpecifier[]::new);
+    }
+
+    private static <T> OrderSpecifier toOrderSpecifier(final EntityPathBase<T> qClass, final Sort.Order sortOrder) {
+        final Order orderMethod = toOrder(sortOrder);
+        final PathBuilder<T> pathBuilder = new PathBuilder<>(qClass.getType(), qClass.getMetadata());
+        return new OrderSpecifier(orderMethod, pathBuilder.get(sortOrder.getProperty()));
+    }
+
+    private static Order toOrder(final Sort.Order sortOrder) {
+        if (sortOrder.isAscending()) {
+            return Order.ASC;
+        }
+        return Order.DESC;
+    }
+}
